Move ExceptionHander under Node namespace
