+++ src/Node/ExceptionHandler.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Phlow\Model;
+namespace Phlow\Node;
 
 /**
- * Interface ExceptionHandlerNode
+ * Interface ExceptionHandler
  * Indicates a Workflow Node that can consume and handle Exceptions thrown during Workflow execution.
  * @package Phlow\Model
  */
-interface ExceptionHandlerNode
+interface ExceptionHandler
 {
     /**
      * Adds a new Exception class that can be handled by this Node
